Mv installer-specific functions to install. Create InstallerStorage.
Two major things happen here:

* Create InstallerStorage is created in blivet.osinstall
This is a subclass of blivet.blivet.Blivet, except it also contains
some installer-specific attrs and functions. Speaking of which--

* Mv installer-specific functions and properties from Blivet ->
InstallerStorage. That list is:
    empty_devices
    unused_devices (prop)
    should_clear
    clear_partitions
    format_by_default
    must_format
    turn_on_swap
    mount_filesystems
    umount_filesystems
    parse_fstab
    mk_dev_root
    create_swap_file
    make_mtab
    add_fstab_swap
    remove_fstab_swap
    set_fstab_swap
    autopart_type (attr)
These were all removed from Blivet and are now in InstallerStorage.

Blivet.do_it was very stripped down; all the installer-specific logic
was moved into InstallerStorage.do_it.
